+++ compile/index.js
@@ -1,22 +1,22 @@
-const { _myNewPackage } = require('./mnp')
+const { _httpErrors } = require('./http-errors')
 
 /**
- * {{ description }}
- * @param {!_myNewPackage.Config} config Options for the program.
+ * Creates HTTP Errors For Goa Apps.
+ * @param {!_httpErrors.Config} config Options for the program.
  * @param {boolean} [config.shouldRun=true] A boolean option. Default `true`.
  * @param {string} [config.text] A text to return.
  * @return {Promise<string>}
  */
-function myNewPackage(config) {
-  return _myNewPackage(config)
+function httpErrors(config) {
+  return _httpErrors(config)
 }
 
-module.exports = myNewPackage
+module.exports = httpErrors
 
 /* typal types/index.xml namespace */
 /**
- * @typedef {_myNewPackage.Config} Config `＠record` Options for the program.
- * @typedef {Object} _myNewPackage.Config `＠record` Options for the program.
+ * @typedef {_httpErrors.Config} Config `＠record` Options for the program.
+ * @typedef {Object} _httpErrors.Config `＠record` Options for the program.
  * @prop {boolean} [shouldRun=true] A boolean option. Default `true`.
  * @prop {string} [text] A text to return.
  */