+++ types/externs.js
@@ -5,19 +5,19 @@
 
 /* typal types/index.xml externs */
 /** @const */
-var _myNewPackage = {}
+var _httpErrors = {}
 /**
  * Options for the program.
  * @record
  */
-_myNewPackage.Config
+_httpErrors.Config
 /**
  * A boolean option. Default `true`.
  * @type {boolean|undefined}
  */
-_myNewPackage.Config.prototype.shouldRun
+_httpErrors.Config.prototype.shouldRun
 /**
  * A text to return.
  * @type {string|undefined}
  */
-_myNewPackage.Config.prototype.text
+_httpErrors.Config.prototype.text